fix: version operations delegate to caller to show error toast
