+++ packages/sanity/src/core/releases/hooks/useVersionOperations.tsx
@@ -1,7 +1,5 @@
 import {useTelemetry} from '@sanity/telemetry/react'
-import {useToast} from '@sanity/ui'
 
-import {useTranslation} from '../../i18n'
 import {type ReleaseId} from '../../perspective/types'
 import {useSetPerspective} from '../../perspective/useSetPerspective'
 import {getDocumentVariantType} from '../../util/getDocumentVariantType'
@@ -25,56 +23,24 @@ export function useVersionOperations(): VersionOperationsValue {
 
   const setPerspective = useSetPerspective()
 
-  const toast = useToast()
-  const {t} = useTranslation()
-
   const handleCreateVersion = async (
     releaseId: ReleaseId,
     documentId: string,
     initialValue?: Record<string, unknown>,
   ) => {
     const origin = getDocumentVariantType(documentId)
-    try {
-      await createVersion(releaseId, documentId, initialValue)
-      setPerspective(releaseId)
-      telemetry.log(AddedVersion, {
-        documentOrigin: origin,
-      })
-    } catch (err) {
-      toast.push({
-        closable: true,
-        status: 'error',
-        title: t('release.action.create-version.failure'),
-        description: err.message,
-      })
-    }
+    await createVersion(releaseId, documentId, initialValue)
+    setPerspective(releaseId)
+    telemetry.log(AddedVersion, {
+      documentOrigin: origin,
+    })
   }
 
-  const handleDiscardVersion = async (releaseId: string, documentId: string) => {
-    try {
-      await discardVersion(releaseId, documentId)
-    } catch (err) {
-      toast.push({
-        closable: true,
-        status: 'error',
-        title: t('release.action.discard-version.failure'),
-        description: err.message,
-      })
-    }
-  }
+  const handleDiscardVersion = async (releaseId: string, documentId: string) =>
+    discardVersion(releaseId, documentId)
 
-  const handleUnpublishVersion = async (documentId: string) => {
-    try {
-      await unpublishVersion(documentId)
-    } catch (err) {
-      toast.push({
-        closable: true,
-        status: 'error',
-        title: t('release.action.unpublish-version.failure'),
-        description: err.message,
-      })
-    }
-  }
+  const handleUnpublishVersion = async (documentId: string) => unpublishVersion(documentId)
+
   return {
     createVersion: handleCreateVersion,
     discardVersion: handleDiscardVersion,
